Refactor voting and pauta services into distinct responsibilities
Split VotacaoService into smaller, specialized services: PautaManagementService, VoteService, and VotingSessionService. Reorganized controllers accordingly for better structure. Enhanced VotingResultResponse and removed redundant methods for streamlined functionality.
