+++ src/main/java/com/votacao/desafio/controller/PautaController.java
@@ -2,22 +2,22 @@ package com.votacao.desafio.controller;
 
 import com.votacao.desafio.dto.PautaRequest;
 import com.votacao.desafio.dto.PautaResponse;
-import com.votacao.desafio.service.PautaService;
+import com.votacao.desafio.service.PautaManagementService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("pauta")
+@RequestMapping("pautas")
 @RequiredArgsConstructor
 public class PautaController {
 
-    private final PautaService pautaService;
+    private final PautaManagementService pautaManagementService;
 
     @PostMapping()
     public ResponseEntity<PautaResponse> createPauta(@RequestBody PautaRequest pautaRequest) {
-        PautaResponse pautaResponse = pautaService.createPauta(pautaRequest);
+        PautaResponse pautaResponse = pautaManagementService.createPauta(pautaRequest);
         return ResponseEntity.ok(pautaResponse);
     }
 
@@ -26,13 +26,13 @@ public class PautaController {
             @RequestParam(required = false) String status,
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "10") int size) {
-        Page<PautaResponse> pautas = pautaService.listAllPautas(status, page, size);
+        Page<PautaResponse> pautas = pautaManagementService.listAllPautas(status, page, size);
         return ResponseEntity.ok(pautas);
     }
 
     @GetMapping("{pautaId}")
     public ResponseEntity<PautaResponse> getPautaById(@PathVariable Long pautaId) {
-        PautaResponse pauta = pautaService.getPautaResponseById(pautaId);
+        PautaResponse pauta = pautaManagementService.getPautaResponseById(pautaId);
         return ResponseEntity.ok(pauta);
     }
 }
